Implemented beamspot radius
Added new option to allow user to adjust radius of
pencil beams. Use

/source/spot R

where R is the radius of the beam in mm. The particle
generator takes into account the direction of the
beam such that generated particles will lie inside
a cylinder of radius R with momentum parallel to
the axis of the cylinder.