Modified the way PMT quantum efficiency is handled
Previously, the quantum efficiency of the PMT was
used simply as a probability whether a photon with
a given wavelength was detected or not. It has been
changed to now modify the amplitude of the photon
signal in the resultant PMT light pulse.

This has the two-fold benefit of both increasing
the number of photon statistics for the same sim
time and resolving the issue of the pulse QDC
being discretized by number of photons detected.