Fixed neutron scatter initialization error
Fixed a bug introduced in commit 278a54a where, in
some cases, the neutron scattering tracking would
not be started correctly. This error arose because
the program was only considering stepping events
where the neutron was at a geometry boundary AND
was within the scintillator material.

This error leads to situations where a significant
number of neutron scatter events are not tracked
correctly because their first scatter event is
within the first column of a segemented detector.
It also caused single-segment detectors to not work
at all because it would never track neutron scatters
in the scintillator body.

This error has been fixed, as far as I can tell.
I've tested both with multi and single segment
detectors and both now work as expected. This also
seems to fix the broken neutron enter positions
nEnterPosX, Y, and Z which have been broken for
some time.