Added kinematic reactions to particle source
/nDet/source/reaction filename.rxn

Where the contents of "filename.rxn" are:

BeamZ	BeamA	BeamBE/A
TargZ	TargA	TargBE/A
EjectZ	EjectA	EjectBE/A	EjectExcitation
RecoilZ	RecoilA	RecoilBE/A	RecoilExcitation
BeamEnergy

Entries must be tab-delimited. Z and A must be
integers and binding energies, excitations, and
beam energies are in MeV.

e.g. for p(t,3He)n with 2.2 MeV protons:

1	1	0
1	3	2.827265
2	3	2.57268	0
0	1	0	0
2.2