Added psuedo-realistic target E-loss effects
Added options for the user to set the thickness of
the target for running simulations with the reaction
particle generator. The user must set both the target
thickness as well as the slope (dE/dx) across the
entire target. The syntax for the reaction command
is given as

/nDet/source/reaction <filename> [thickness] [dE/dx]

where the thickness is given in (mm) and dE/dx is
given in (MeV/mm). The slope dE/dx should always
be negative but there are no restrictions on the
input value.