Updates for multi-detector output
Moved centerOfMass objects from nDetRunAction
to userAddDetector class since that is now where
all the detector information is stored.

Added master thread to nDetThreadContainer class.
This is done so that we can update the list of
detectors for the master thread.

Moved pmt response settings from nDetRunAction to
nDetConstruction.

Added set-able digitization/integral settings to
pmtResponse class so that the caller will not need
to supply these values.

TODO:
  G4GeneralParticleSource causes seg-faults when
  the code is run in multithreading mode. These
  faults are caused by calls to ClearAll()