Added new detector macro commands
Added new macro commands to set the width and height
of scintillator segments for segmented detectors.
These may be used to set a constant size for the
segment width and height without manually setting
the detector width or height. The syntax for these
commands is:

/nDet/detector/setSegmentWidth <width(mm)>
/nDet/detector/setSegmentHeight <height(mm)>

Added new command to print information about all
currently defined detectors. The syntax is

/nDet/detector/print