Added new commands to nDetDynamicMaterial
Added 'setDebug' command to nDetDynamicMaterial
class which can be used to toggle the dynamic
material builder debug mode.

Added 'setMaterial' command to nDetDynamicMaterial
class which can be used to copy a material from
the NIST database instead of building one from
scratch. This is useful in the event the user
wants to use a pre-defined material but to modify
the physical parameters slightly.