Modified the way the light response pulse is built
Changed the way the optical photon response pulse is
generated by the pmtResponse class.

The response pulse is generated by stepping through
time (in increments of the ADC clock period). At each
step, the total response is sampled by summing the
contributing response from every detected photon.

I believe the cause of the bias in the light pulse
timing was due to the way I selected the bin to
start computing the single-photon response (line
number 203 in pmtResponse.cc):

size_t index = (size_t)floor(dt/ADC_CLOCK_TICK);

This index is no longer used in the new summing
method.

Added ADC latching to pmtResponse class. The ADC
latch is used to set a random time offset between the
ADC clock and the input signal when it is digitized.