Moved detector center-of-mass copying
Moved where the center-of-mass calculators are
copied to the defined detectors. Previously, they
were copied whenever UpdateGeometry() was called. This
has been moved to immediately after the new detector
is created in AddGeometry(). This allows the user
more flexibility because they may now change the
pmt segmentation, gain matrix, or spectral response
for each new detector.