Copies of all detectors are cloned for each thread
nDetConstruction class now copies all detectors
for the run action of each thread. This ensures
that each thread will have its own instance of
every detector when running in multithreaded
mode.

Implemented clone methods for nDetDetector,
centerOfMass, and pmtResponse classes to ensure
that their variable fields are set exactly when
they are copied.

Removed auto for-loops from nDetRunAction since
the vector of detectors is now a vector of objects
instead of a vector of pointers. The auto loops
were causing issues due to the creation of local
copies of the detectors inside the for loops.