Added changelog.sh shell script
Added shell script changelog.sh which may be used
to easily generate a new NEXTSim changelog entry
which conforms to the format which is present in
the NEXTSim wiki changelog.

When merging dev commits into the master branch, you
should add an entry to the changelog so that regular
users know what changes have been made to the source.

To create a new master release:

  git checkout master
  git merge dev
  ./cmake/changelog.sh PREVIOUS_COMMIT

where PREVIOUS_COMMIT is the first 7 characters of
the most recently merged master commit e.g. 28ada18
(you can obtain this number from the changelog page
in the NEXTSim wiki, it will be the top-most entry
in the version table).

The script will output the new merge documentation
and will direct you where to add the information to
the changelog.md file in the NEXTSim wiki.